custom-elements/instantiating-custom-elements/custom-element-type-local-name-and-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type is the local name of the custom element even if IS attribute provided</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If both types of custom element types are provided at the time of element's instantiation, the custom tag must win over the type extension">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    doc.registerElement('x-a-1');
    doc.registerElement('x-a-2');
    doc.body.innerHTML = '<x-a-1 id="xa" is="x-a-2"></x-a-1>';
    var customElement = doc.querySelector('#xa');
    
    assert_equals(customElement.constructor.prototype.constructor.name, 
            customElement.localName, 'Wrong custom element local name');
    assert_equals(customElement.constructor.prototype.constructor.name, 'x-a-1', 
        'Wrong custom element type');    
}, 'Custom element type is the local name of the custom element even if IS attribute provided.');


test(function() {
    var doc = newHTMLDocument();
    
    doc.registerElement('x-a-4');
    doc.body.innerHTML = '<x-a-3 id="xa" is="x-a-4"></x-a-3>';
    var customElement = doc.querySelector('#xa');
    assert_equals(customElement.constructor.prototype.constructor.name, 'HTMLElement', 
        'Wrong custom element local name');

    doc.registerElement('x-a-3');
    assert_equals(customElement.constructor.prototype.constructor.name, 
            customElement.localName, 'Wrong custom element local name');
    assert_equals(customElement.constructor.prototype.constructor.name, 'x-a-3', 
            'Wrong custom element type');
}, 'Custom element type is the local name of the custom element even if IS attribute provided.' + 
    'Custom element is unresolved at first');


test(function() {
    var doc = newHTMLDocument();
    doc.registerElement('x-element');
    
    HTML5_ELEMENTS.forEach(function(value) {
        var obj = doc.createElement(value);
        var name = 'x-' + value + '-01';
        doc.registerElement(name, {
            prototype: Object.create(obj.constructor.prototype),
            extends: value
        });
        doc.body.innerHTML = '<x-element id="xelement" is="' + name + '"></x-element>';
        var customElement = doc.querySelector('#xelement');
        assert_equals(customElement.constructor.name, 
                customElement.localName, 'Wrong custom element local name');
        assert_equals(customElement.constructor.name, 'x-element', 
            'Wrong custom element type');            
    });
}, 'Custom element type is the local name of the custom element even if IS attribute provided.' +
    'Test extension of all HTML5 elements');


test(function() {
    var doc = newHTMLDocument();
    doc.registerElement('x-element-2');
    
    HTML5_ELEMENTS.forEach(function(value) {
        var obj = doc.createElement(value);
        var name = 'x-' + value + '-02';
        var id = 'x_' + value + '_02';
        doc.registerElement(name, {
            prototype: Object.create(obj.constructor.prototype),
            extends: value
        });
        doc.body.innerHTML = '<' + name + ' id="' + id + '" is="x-element-2"></' + name + '>';
        var customElement = doc.querySelector('#' + id);
        // We have <x-a-02 is='x-element-2'>. Custom element type for this will be
        // HTMLElement, not x-a-02 (for x-a-02 there should be <a is-x-a-02>...) 
        assert_equals(customElement.constructor.name, 'HTMLElement', 
            'Wrong custom element type');            
    });
}, 'Custom element type is the local name of the custom element even if IS attribute provided.' +
    'Another test of extension of all HTML5 elements.');

</script>
</body>
</html>
